added new CaseRoleTest file, renamed old file
