+++ src/test/java/uk/gov/hmcts/reform/roleassignment/domain/service/drools/DroolJudicialCategoryCaseRoleTest.java
@@ -1,0 +1,52 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.drools;
+
+import org.junit.jupiter.api.Test;
+import org.junit.runner.RunWith;
+import org.mockito.junit.MockitoJUnitRunner;
+import uk.gov.hmcts.reform.roleassignment.domain.model.RoleAssignment;
+import uk.gov.hmcts.reform.roleassignment.domain.model.enums.GrantType;
+import uk.gov.hmcts.reform.roleassignment.domain.model.enums.RoleCategory;
+import uk.gov.hmcts.reform.roleassignment.domain.model.enums.RoleType;
+import uk.gov.hmcts.reform.roleassignment.domain.model.enums.Status;
+
+import java.time.ZoneOffset;
+import java.time.ZonedDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static uk.gov.hmcts.reform.roleassignment.domain.model.enums.GrantType.STANDARD;
+import static uk.gov.hmcts.reform.roleassignment.helper.TestDataBuilder.getRequestedOrgRole;
+import static uk.gov.hmcts.reform.roleassignment.util.JacksonUtils.convertValueJsonNode;
+
+@RunWith(MockitoJUnitRunner.class)
+class DroolJudicialCategoryCaseRoleTest extends DroolBase {
+
+    @Test
+    void shouldApproveRequestedRoleForCase() {
+
+        RoleAssignment requestedRole1 =  getRequestedCaseRole(RoleCategory.JUDICIAL, "judge", GrantType.SPECIFIC);
+        requestedRole1.getAttributes().put("caseId", convertValueJsonNode("1234567890123456"));
+
+        List<RoleAssignment> requestedRoles = new ArrayList<>();
+        requestedRoles.add(requestedRole1);
+
+        assignmentRequest.setRequestedRoles(requestedRoles);
+        assignmentRequest.getRequestedRoles().stream().forEach(roleAssignment -> {
+
+            roleAssignment.setBeginTime(ZonedDateTime.now(ZoneOffset.UTC));
+        });
+
+        //Execute Kie session
+        buildExecuteKieSession();
+
+        //assertion
+        assignmentRequest.getRequestedRoles().stream().forEach(roleAssignment ->
+                                                                   assertEquals(
+                                                                       Status.APPROVED,
+                                                                       roleAssignment.getStatus()
+                                                                   )
+        );
+    }
+
+}
