+++ src/test/java/uk/gov/hmcts/reform/roleassignment/domain/service/drools/DroolJudicialCategoryOrgRoleTest.java
@@ -20,34 +20,7 @@ import static uk.gov.hmcts.reform.roleassignment.helper.TestDataBuilder.getReque
 import static uk.gov.hmcts.reform.roleassignment.util.JacksonUtils.convertValueJsonNode;
 
 @RunWith(MockitoJUnitRunner.class)
-class DroolJudicialCategoryTest extends DroolBase {
-
-    @Test
-    void shouldApproveRequestedRoleForCase() {
-
-        RoleAssignment requestedRole1 =  getRequestedCaseRole(RoleCategory.JUDICIAL, "judge", GrantType.SPECIFIC);
-        requestedRole1.getAttributes().put("caseId", convertValueJsonNode("1234567890123456"));
-
-        List<RoleAssignment> requestedRoles = new ArrayList<>();
-        requestedRoles.add(requestedRole1);
-
-        assignmentRequest.setRequestedRoles(requestedRoles);
-        assignmentRequest.getRequestedRoles().stream().forEach(roleAssignment -> {
-
-            roleAssignment.setBeginTime(ZonedDateTime.now(ZoneOffset.UTC));
-        });
-
-        //Execute Kie session
-        buildExecuteKieSession();
-
-        //assertion
-        assignmentRequest.getRequestedRoles().stream().forEach(roleAssignment ->
-                                                                   assertEquals(
-                                                                       Status.APPROVED,
-                                                                       roleAssignment.getStatus()
-                                                                   )
-        );
-    }
+class DroolJudicialCategoryOrgRoleTest extends DroolBase {
 
     @Test
     void shouldRejectCaseValidationForRequestedRoleMisssingCaseId() {
